medbuddy/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
google-generativeai==0.7.2
twilio==9.0.5
apscheduler==3.10.4
python-dotenv==1.0.1
Pillow==12.1.1
aiofiles==23.2.1